pyproject.toml
[project]
name = "toetsen-schouwen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openpyxl>=3.1.5",
  "pandas>=2.3.2",
  "pytest>=8.4.1",
  "python-dotenv>=1.1.1",
  "requests>=2.32.5",
]

[tool.pytest.ini_options]
addopts = ["--strict-markers"]
markers = [
  "deel_4",
  "deel_7",
  "deel_9",
  "deel_4_reset",
  "deel_7_reset",
  "deel_9_reset",
  "deel_4_verwacht",
  "deel_7_verwacht",
  "deel_9_verwacht",
]

log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(message)s"

log_file = "pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
